指標出力のテストコード追加
