+++ tests/test_classifier/output_dir/indicator.json
@@ -1,0 +1,28 @@
+{
+  "accuracy": 0.9385964912280702,
+  "precision": 0.9384615384615385,
+  "recall": 0.953125,
+  "confusion_matrix": {
+    "TN": 46,
+    "FP": 4,
+    "FN": 3,
+    "TP": 61
+  },
+  "auc": 0.91890625,
+  "roc": {
+    "fpr": [
+      0.0,
+      0.04,
+      0.08,
+      0.1,
+      1.0
+    ],
+    "tpr": [
+      0.0,
+      0.046875,
+      0.953125,
+      0.953125,
+      1.0
+    ]
+  }
+}
